Remove NFT and old MarketOverview code
