--- app/src/main/java/io/horizontalsystems/bankwallet/modules/market/overview/ui/MarketOverviewBoards.kt
@@ -1,123 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.market.overview.ui
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.LocalMinimumInteractiveComponentEnforcement
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavController
-import io.horizontalsystems.bankwallet.R
-import io.horizontalsystems.bankwallet.core.slideFromRight
-import io.horizontalsystems.bankwallet.core.stats.StatEvent
-import io.horizontalsystems.bankwallet.core.stats.StatPage
-import io.horizontalsystems.bankwallet.core.stats.stat
-import io.horizontalsystems.bankwallet.core.stats.statSection
-import io.horizontalsystems.bankwallet.modules.coin.CoinFragment
-import io.horizontalsystems.bankwallet.modules.market.MarketModule
-import io.horizontalsystems.bankwallet.modules.market.MarketViewItem
-import io.horizontalsystems.bankwallet.modules.market.TopMarket
-import io.horizontalsystems.bankwallet.modules.market.overview.MarketOverviewModule
-import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
-import io.horizontalsystems.bankwallet.ui.compose.Select
-import io.horizontalsystems.bankwallet.ui.compose.WithTranslatableTitle
-import io.horizontalsystems.bankwallet.ui.compose.components.ButtonSecondaryToggle
-import io.horizontalsystems.bankwallet.ui.compose.components.MarketCoinClear
-
-@Composable
-fun BoardsView(
-    boards: List<MarketOverviewModule.Board>,
-    navController: NavController,
-    onClickSeeAll: (MarketModule.ListType) -> Unit,
-    onSelectTopMarket: (TopMarket, MarketModule.ListType) -> Unit
-) {
-    val onItemClick: (MarketViewItem) -> Unit = remember {
-        {
-            navController.slideFromRight(
-                R.id.coinFragment,
-                CoinFragment.Input(it.coinUid)
-            )
-        }
-    }
-
-    boards.forEach { boardItem ->
-        TopBoardHeader(
-            title = boardItem.boardHeader.title,
-            iconRes = boardItem.boardHeader.iconRes,
-            select = boardItem.boardHeader.topMarketSelect,
-            onSelect = { topMarket -> onSelectTopMarket(topMarket, boardItem.type) },
-            onClickSeeAll = { onClickSeeAll(boardItem.type) }
-        )
-
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 16.dp)
-                .clip(RoundedCornerShape(12.dp))
-                .background(ComposeAppTheme.colors.lawrence)
-        ){
-            boardItem.marketViewItems.forEach { coin ->
-                MarketCoinWithBackground(coin) {
-                    onItemClick.invoke(coin)
-
-                    stat(page = StatPage.MarketOverview, section = boardItem.type.statSection, event = StatEvent.OpenCoin(coin.coinUid))
-                }
-            }
-
-            SeeAllButton { onClickSeeAll(boardItem.type) }
-        }
-
-        Spacer(modifier = Modifier.height(24.dp))
-    }
-}
-
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
-fun <T : WithTranslatableTitle> TopBoardHeader(
-    title: Int,
-    iconRes: Int,
-    select: Select<T>,
-    onSelect: (T) -> Unit,
-    onClickSeeAll: () -> Unit
-) {
-    MarketsSectionHeader(
-        title = title,
-        onClick = onClickSeeAll,
-        icon = painterResource(iconRes)
-    ) {
-        CompositionLocalProvider(LocalMinimumInteractiveComponentEnforcement provides false) {
-            ButtonSecondaryToggle(
-                select = select,
-                onSelect = onSelect
-            )
-        }
-    }
-}
-
-@Composable
-private fun MarketCoinWithBackground(
-    marketViewItem: MarketViewItem,
-    onClick: () -> Unit
-) {
-    MarketCoinClear(
-        marketViewItem.coinCode,
-        marketViewItem.coinName,
-        marketViewItem.iconUrl,
-        marketViewItem.alternativeIconUrl,
-        marketViewItem.iconPlaceHolder,
-        marketViewItem.value,
-        marketViewItem.marketDataValue,
-        marketViewItem.rank,
-        onClick
-    )
-}
