--- app/src/main/java/io/horizontalsystems/bankwallet/modules/market/overview/ui/MarketOverviewTopPlatforms.kt
@@ -1,50 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.market.overview.ui
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.unit.dp
-import io.horizontalsystems.bankwallet.modules.market.TimeDuration
-import io.horizontalsystems.bankwallet.modules.market.overview.MarketOverviewModule
-import io.horizontalsystems.bankwallet.modules.market.topplatforms.Platform
-import io.horizontalsystems.bankwallet.modules.market.topplatforms.TopPlatformItem
-import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
-
-@Composable
-fun TopPlatformsBoardView(
-    board: MarketOverviewModule.TopPlatformsBoard,
-    onSelectTimeDuration: (TimeDuration) -> Unit,
-    onItemClick: (Platform) -> Unit,
-    onClickSeeAll: () -> Unit
-) {
-    TopBoardHeader(
-        title = board.title,
-        iconRes = board.iconRes,
-        select = board.timeDurationSelect,
-        onSelect = onSelectTimeDuration,
-        onClickSeeAll = onClickSeeAll
-    )
-
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp)
-            .clip(RoundedCornerShape(12.dp))
-            .background(ComposeAppTheme.colors.lawrence)
-    ) {
-        board.items.forEach {
-            TopPlatformItem(it, onItemClick)
-        }
-
-        SeeAllButton(onClickSeeAll)
-    }
-
-    Spacer(modifier = Modifier.height(24.dp))
-}
