--- app/src/main/java/io/horizontalsystems/bankwallet/modules/nft/collection/assets/NftCollectionAssetsModule.kt
@@ -1,27 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.nft.collection.assets
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import io.horizontalsystems.bankwallet.core.App
-import io.horizontalsystems.bankwallet.modules.balance.BalanceXRateRepository
-import io.horizontalsystems.marketkit.models.BlockchainType
-
-object NftCollectionAssetsModule {
-
-    class Factory(
-        private val blockchainType: BlockchainType,
-        private val collectionUid: String
-    ) : ViewModelProvider.Factory {
-        @Suppress("UNCHECKED_CAST")
-        override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            val service = NftCollectionAssetsService(
-                blockchainType,
-                collectionUid,
-                App.nftMetadataManager.provider(blockchainType),
-                BalanceXRateRepository("nft-collection-assets", App.currencyManager, App.marketKit)
-            )
-            return NftCollectionAssetsViewModel(service) as T
-        }
-    }
-
-}
