--- app/src/main/java/io/horizontalsystems/bankwallet/modules/nft/collection/events/NftCollectionEventsModule.kt
@@ -1,44 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.nft.collection.events
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import io.horizontalsystems.bankwallet.core.App
-import io.horizontalsystems.bankwallet.core.providers.nft.NftEventsProvider
-import io.horizontalsystems.bankwallet.entities.nft.NftEventMetadata
-import io.horizontalsystems.bankwallet.entities.nft.NftUid
-import io.horizontalsystems.bankwallet.modules.balance.BalanceXRateRepository
-import io.horizontalsystems.bankwallet.modules.coin.ContractInfo
-import io.horizontalsystems.marketkit.models.BlockchainType
-
-class NftCollectionEventsModule {
-
-    class Factory(
-        private val eventListType: NftEventListType,
-        private val defaultEventType: NftEventMetadata.EventType = NftEventMetadata.EventType.Sale
-    ) : ViewModelProvider.Factory {
-        @Suppress("UNCHECKED_CAST")
-        override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            val service = NftCollectionEventsService(
-                eventListType,
-                defaultEventType,
-                NftEventsProvider(App.marketKit),
-                BalanceXRateRepository("nft-collection-events", App.currencyManager, App.marketKit)
-            )
-            return NftCollectionEventsViewModel(service) as T
-        }
-    }
-}
-
-enum class SelectorDialogState {
-     Closed, Opened
-}
-
-sealed class NftEventListType {
-    data class Collection(
-        val blockchainType: BlockchainType,
-        val providerUid: String,
-        val contracts: List<ContractInfo>
-    ) : NftEventListType()
-
-    data class Asset(val nftUid: NftUid) : NftEventListType()
-}
