--- app/src/main/java/io/horizontalsystems/bankwallet/modules/nft/ui/CellLink.kt
@@ -1,44 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.nft.ui
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Icon
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import io.horizontalsystems.bankwallet.R
-import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
-import io.horizontalsystems.bankwallet.ui.compose.components.body_leah
-
-@Composable
-fun CellLink(icon: Painter, title: String, onClick: () -> Unit) {
-    Row(
-        modifier = Modifier
-            .fillMaxSize()
-            .clickable(onClick = onClick),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-
-        Icon(
-            modifier = Modifier.padding(horizontal = 16.dp),
-            painter = icon,
-            contentDescription = null,
-            tint = ComposeAppTheme.colors.grey
-        )
-        body_leah(
-            modifier = Modifier.weight(1f),
-            text = title,
-        )
-        Icon(
-            modifier = Modifier.padding(horizontal = 16.dp),
-            painter = painterResource(id = R.drawable.ic_arrow_right),
-            contentDescription = null,
-            tint = ComposeAppTheme.colors.grey
-        )
-    }
-}
