+++ app/src/main/java/io/horizontalsystems/bankwallet/widgets/MarketWidgetManager.kt
@@ -150,8 +150,6 @@ class MarketWidgetManager {
     companion object {
         fun getMarketWidgetTypes(): List<MarketWidgetType> {
             val types = MarketWidgetType.values().toMutableList()
-            // TopNfts type is hidden for now. It will be removed in next sprints
-            types.remove(MarketWidgetType.TopNfts)
             if (!App.localStorage.marketsTabEnabled) {
                 types.remove(MarketWidgetType.Watchlist)
             }
