+++ app/src/main/java/io/horizontalsystems/bankwallet/widgets/MarketWidgetRepository.kt
@@ -16,8 +16,6 @@ import io.horizontalsystems.bankwallet.modules.market.favorites.MarketFavoritesM
 import io.horizontalsystems.bankwallet.modules.market.favorites.WatchlistSorting
 import io.horizontalsystems.bankwallet.modules.market.favorites.period
 import io.horizontalsystems.bankwallet.modules.market.sort
-import io.horizontalsystems.bankwallet.modules.market.topnftcollections.TopNftCollectionsRepository
-import io.horizontalsystems.bankwallet.modules.market.topnftcollections.TopNftCollectionsViewItemFactory
 import io.horizontalsystems.bankwallet.modules.market.topplatforms.TopPlatformsRepository
 import kotlinx.coroutines.rx2.await
 
@@ -25,8 +23,6 @@ class MarketWidgetRepository(
     private val marketKit: MarketKitWrapper,
     private val favoritesManager: MarketFavoritesManager,
     private val favoritesMenuService: MarketFavoritesMenuService,
-    private val topNftCollectionsRepository: TopNftCollectionsRepository,
-    private val topNftCollectionsViewItemFactory: TopNftCollectionsViewItemFactory,
     private val topPlatformsRepository: TopPlatformsRepository,
     private val currencyManager: CurrencyManager
 ) {
@@ -50,10 +46,6 @@ class MarketWidgetRepository(
 
             MarketWidgetType.TopGainers -> {
                 getTopGainers()
-            }
-
-            MarketWidgetType.TopNfts -> {
-                getTopNtfs()
             }
 
             MarketWidgetType.TopPlatforms -> {
@@ -86,30 +78,6 @@ class MarketWidgetRepository(
                 diff = item.changeDiff,
                 blockchainTypeUid = null,
                 imageRemoteUrl = item.platform.iconUrl
-            )
-        }
-    }
-
-    private suspend fun getTopNtfs(): List<MarketWidgetItem> {
-        val nftCollectionViewItems = topNftCollectionsRepository.get(
-            sortingField = SortingField.HighestVolume,
-            timeDuration = TimeDuration.SevenDay,
-            forceRefresh = true,
-            limit = itemsLimit
-        ).mapIndexed { index, item ->
-            topNftCollectionsViewItemFactory.viewItem(item, TimeDuration.SevenDay, index + 1)
-        }
-
-        return nftCollectionViewItems.map {
-            MarketWidgetItem(
-                uid = it.uid,
-                title = it.name,
-                subtitle = it.floorPrice,
-                label = it.order.toString(),
-                value = it.volume,
-                diff = it.volumeDiff,
-                blockchainTypeUid = it.blockchainType.uid,
-                imageRemoteUrl = it.imageUrl ?: ""
             )
         }
     }
