+++ app/src/main/java/io/horizontalsystems/bankwallet/widgets/MarketWidgetState.kt
@@ -27,7 +27,6 @@ data class MarketWidgetState(
 enum class MarketWidgetType(val title: Int, val id: String) {
     Watchlist(R.string.Market_Tab_Watchlist, "watchlist"),
     TopGainers(R.string.RateList_TopGainers, "topGainers"),
-    TopNfts(R.string.Nft_TopCollections, "topNfts"),
     TopPlatforms(R.string.MarketTopPlatforms_Title, "topPlatforms");
 
     companion object {
